pyproject.toml
[tool.poetry]
name = "text_classification_transformers"
version = "0.1.0"
description = "text classification samples with transformers"
authors = ["i13harada1s"]
license = "MIT"

[tool.isort]
force_sort_within_sections = true
profile = "black"

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.21.0"
datasets = "^2.4.0"
absl-py = "^1.2.0"
fugashi = {extras = ["unidic-lite"], version = "^1.1.2"}
torch = {url = "https://download.pytorch.org/whl/cu113/torch-1.10.1%2Bcu113-cp39-cp39-linux_x86_64.whl"}
scikit-learn = "^1.1.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
